linux: install perf in development builds
perf will be built if CONFIG_PERF_EVENTS is enabled in the
kernel config.

elfutils, binutils and libunwind are needed to get symbols
of userspace programs in perf output.

zlib is an implicit dependency of elfutils and and optional
one for perf - but better add it here.

openssl/libcrypto is an optional dependency for perf, add it
so that the perf configuration on clean build is identical
to an unclean rebuild.

Signed-off-by: Matthias Reichl <hias@horus.com>