RPi, Generic: backport IR latency patch and mce_kbd/mceusb fixes
See https://www.spinics.net/lists/linux-media/msg131804.html
https://patchwork.linuxtv.org/patch/48681/
https://patchwork.linuxtv.org/patch/48680/
https://patchwork.linuxtv.org/patch/48782/

The current IR decoding is much too slow. Many IR protocols rely on
a trailing space for decoding (e.g. rc-6 needs to know when the bits
end). The trailing space is generated by the IR timeout, and if this
is longer than required, buttons can feel slow to respond.

The other issue is the keyup timer. IR has no concept of a keyup message,
this is implied by the absence of IR. So, minimising the timeout for
this makes buttons less "sticky"; the are released much quicker.

With these patches in place, using IR with the builtin decoders is much
improved and feels very snappy.

Signed-off-by: Matthias Reichl <hias@horus.com>