rtk_hciattach: don't read mac from vendor storage
This tool was designed for use on rockchip systems, it attempts to
read the mac from /dev/vendor_storage and when it can't access it
then it generates a new random mac on each use of it, this in turn
leads to users having to re-pair their BT devices on each reboot.

Removing the define causes the tool to use the predefined mac which
is better than a random mac.

Also read BT mac from /storage/.config/btaddr if it exists.